extract [nat] to [Int] rather than [Integer]
as a very rough estimate, our performance seems to be within ~10x of what
we should expect for our design.  on linux, a vaguely similar benchmark
(grow-write-read-shrink on an existing file descriptor) takes 0.07s,
in contrast to our 1.6s, but we are doing more (creating log entries
for two transactions in each iteration, in contrast to linux updating
the file cache).  perhaps the xv6 file system would be another good
baseline for performance comparisons.

+++ src/hslib/Word.hs
@@ -8,47 +8,47 @@ data Coq_word = W !Integer
 bitmask :: [Integer]
 bitmask = map (\x -> 2^x - 1) [(0::Int) ..]
 
-wrap :: Integer -> Integer -> Integer
-wrap nbits v = (Data.Bits..&.) v (bitmask !! fromIntegral nbits)
+wrap :: Int -> Integer -> Integer
+wrap nbits v = (Data.Bits..&.) v (bitmask !! nbits)
 
-weq :: Integer -> Coq_word -> Coq_word -> Bool
+weq :: Int -> Coq_word -> Coq_word -> Bool
 weq _ (W x) (W y) = x == y
 
-wlt_dec :: Integer -> Coq_word -> Coq_word -> Bool
+wlt_dec :: Int -> Coq_word -> Coq_word -> Bool
 wlt_dec _ (W x) (W y) = x < y
 
-wplus :: Integer -> Coq_word -> Coq_word -> Coq_word
+wplus :: Int -> Coq_word -> Coq_word -> Coq_word
 wplus n (W x) (W y) = W $ wrap n (x + y)
 
-wminus :: Integer -> Coq_word -> Coq_word -> Coq_word
+wminus :: Int -> Coq_word -> Coq_word -> Coq_word
 wminus n (W x) (W y) = W $ wrap n (2^n + x - y)
 
-wmult :: Integer -> Coq_word -> Coq_word -> Coq_word
+wmult :: Int -> Coq_word -> Coq_word -> Coq_word
 wmult n (W x) (W y) = W $ wrap n (x * y)
 
-wdiv :: Integer -> Coq_word -> Coq_word -> Coq_word
+wdiv :: Int -> Coq_word -> Coq_word -> Coq_word
 wdiv _ (W x) (W y) = W $ x `quot` y
 
-wmod :: Integer -> Coq_word -> Coq_word -> Coq_word
+wmod :: Int -> Coq_word -> Coq_word -> Coq_word
 wmod _ (W x) (W y) = W $ x `rem` y
 
-natToWord :: Integer -> Integer -> Coq_word
+natToWord :: Int -> Int -> Coq_word
 natToWord _ 0 = W 0
 natToWord n x = W $ wrap n $ fromIntegral x
 
-wordToNat :: Integer -> Coq_word -> Integer
-wordToNat _ (W x) = x
+wordToNat :: Int -> Coq_word -> Int
+wordToNat _ (W x) = fromIntegral x
 
-zext :: Integer -> Coq_word -> Integer -> Coq_word
+zext :: Int -> Coq_word -> Int -> Coq_word
 zext _ (W w) _ = W w
 
-split1 :: Integer -> Integer -> Coq_word -> Coq_word
+split1 :: Int -> Int -> Coq_word -> Coq_word
 split1 sz1 _ (W w) = W $ wrap sz1 w
 
-split2 :: Integer -> Integer -> Coq_word -> Coq_word
+split2 :: Int -> Int -> Coq_word -> Coq_word
 split2 sz1 _ (W w) = W $ w `Data.Bits.shiftR` (fromIntegral sz1)
 
-combine :: Integer -> Coq_word -> Integer -> Coq_word -> Coq_word
+combine :: Int -> Coq_word -> Int -> Coq_word -> Coq_word
 combine sz1 (W w1) _ (W w2) = W $ w1 + (w2 `Data.Bits.shiftL` (fromIntegral sz1))
 
 instance Show Coq_word where
